Rename shared lib
