+++ libs/shared/util/src/main/java/org/example/awesome/UtilApplication.java
@@ -1,4 +1,4 @@
-package org.example.project;
+package org.example.awesome;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
