+++ libs/shared/util/src/main/resources/application.yml
