+++ libs/shared/util/src/test/java/org/example/awesome/UtilApplicationTests.java
@@ -1,4 +1,4 @@
-package org.example.project;
+package org.example.awesome;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
